requirements.txt
python-dotenv==1.0.1
openai==1.54.4
langchain-openai==0.2.2
langchain-core==0.3.9
